Fix schema

+++ db/migrate/20220114220208_create_bulk_discounts.rb
@@ -2,7 +2,7 @@ class CreateBulkDiscounts < ActiveRecord::Migration[5.2]
   def change
     create_table :bulk_discounts do |t|
       t.string :name
-      t.float :discount
+      t.integer :discount
       t.integer :threshold
       t.references :merchant, foreign_key: true
